# src/app/api/reports/export/route.ts
import { NextResponse } from 'next/server';
import { createClient } from '@/lib/supabase/server';

export async function POST(request: Request) {
  const supabase = createClient();
  const {
    data: { user },
    error,
  } = await supabase.auth.getUser();

  if (error || !user) {
    return NextResponse.json({ error: 'Unauthorized', code: 'UNAUTHORIZED' }, { status: 401 });
  }

  try {
    const body = await request.json();
    const { reportType, format, filters } = body;

    if (!reportType || !format) {
      return NextResponse.json(
        { error: 'Report type and format are required', code: 'MISSING_PARAMS' },
        { status: 400 },
      );
    }

    // For MVP, we'll return a simple message indicating export capability
    // Full implementation with jspdf, csv-writer, exceljs to be completed
    const response = {
      message: `Export functionality for ${format.toUpperCase()} format will be available in the next update`,
      reportType,
      format,
      filters,
      note: 'For now, you can copy the report data from the screen',
    };

    return NextResponse.json(response);
  } catch (error) {
    console.error('Error exporting report:', error);
    return NextResponse.json(
      { error: 'Internal server error', code: 'INTERNAL_ERROR' },
      { status: 500 },
    );
  }
}